Increase typehints, use consistent token unpacking in Transformers

+++ test/functions/builtins/test_background.py
@@ -33,7 +33,7 @@ class TestBackground:
     def test_background(
         self,
         argument: "WaveformVariable | np.ndarray",
-        expected_result: int,
+        expected_result: float,
     ):
         result = Background.evaluate(argument)
